pipeline/requirements.txt
anthropic>=0.28.0
google-genai>=1.0.0
supabase>=2.5.0
youtube-transcript-api>=0.6.2
google-api-python-client>=2.130.0
assemblyai>=0.30.0
feedparser>=6.0.11
requests>=2.32.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.1
schedule>=1.2.2
